pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spatial-grid"
version = "0.1.0"
description = "Generate local geophysical grids for survey and drill-hole planning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Drew Heasman", email = "heasman.drew@gmail.com"}]
keywords = ["geophysics", "exploration", "mining", "GIS", "drilling", "survey"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "geopandas>=0.14",
    "shapely>=2.0",
    "pyproj>=3.6",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "click>=8.1",
    "simplekml>=1.3",
    "gpxpy>=1.6",
]

[project.optional-dependencies]
dev = ["pytest>=7", "ruff>=0.1"]

[project.scripts]
spatial-grid = "spatial_grid.cli:main"

[project.urls]
Homepage = "https://github.com/dheasman/spatial-grid"
Issues = "https://github.com/dheasman/spatial-grid/issues"

[tool.hatch.build.targets.wheel]
packages = ["spatial_grid"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
